refactor: use core/enable_if over utility/enable_if
The later is deprecated.
