+++ include/boost/bimap/relation/detail/metadata_access_builder.hpp
@@ -20,7 +20,7 @@
 
 #include <boost/bimap/relation/support/is_tag_of_member_at.hpp>
 #include <boost/bimap/detail/debug/static_error.hpp>
-#include <boost/utility/enable_if.hpp>
+#include <boost/core/enable_if.hpp>
 #include <boost/preprocessor/cat.hpp>
