+++ include/boost/bimap/relation/detail/mutant.hpp
@@ -25,7 +25,7 @@
 #include <boost/type_traits/is_const.hpp>
 #include <boost/utility/addressof.hpp>
 #include <boost/mpl/not.hpp>
-#include <boost/utility/enable_if.hpp>
+#include <boost/core/enable_if.hpp>
 
 namespace boost {
 namespace bimaps {
