+++ include/boost/bimap/relation/structured_pair.hpp
@@ -26,7 +26,7 @@
 
 #include <boost/call_traits.hpp>
 
-#include <boost/utility/enable_if.hpp>
+#include <boost/core/enable_if.hpp>
 #include <boost/type_traits/is_same.hpp>
 #include <boost/mpl/if.hpp>
 #include <boost/mpl/vector.hpp>
