+++ include/boost/bimap/relation/support/is_tag_of_member_at.hpp
@@ -22,7 +22,7 @@
 #include <boost/type_traits/is_same.hpp>
 
 #include <boost/mpl/bool.hpp>
-#include <boost/utility/enable_if.hpp>
+#include <boost/core/enable_if.hpp>
 #include <boost/bimap/relation/support/member_with_tag.hpp>
 
 /** \struct boost::bimaps::relation::support::is_tag_of_member_at_left
